requirements.txt
alpaca-trade-api
pandas
pandas-ta
requests
python-dotenv
mplfinance
matplotlib
pytz